[fix] #34 - dto명 수정 및 로그확인용 코드 삭제
