+++ src/main/java/gitbal/backend/entity/dto/UserInfoDto.java
@@ -1,12 +1,8 @@
 package gitbal.backend.entity.dto;
 
+public record UserInfoDto2(String userName, String univName, String regionName, String imgName ,String userTitle) {
 
-public record UserInfoDto(Long prCount, Long commitCount, Long issues,
-                          Long followerCount, Long repoCount) {
-    public static UserInfoDto of(Long prCount, Long commitCount, Long issues,
-        Long followerCount, Long repoCount) {
-        return new UserInfoDto(prCount, commitCount, issues, followerCount, repoCount);
-    }
-
-
+  public static UserInfoDto2 of (String userName, String univName, String regionName, String imgName ,String userTitle){
+    return new UserInfoDto2(userName, univName, regionName, imgName, userTitle);
+  }
 }
