--- src/main/java/gitbal/backend/entity/dto/UserInfoDto2.java
@@ -1,8 +1,0 @@
-package gitbal.backend.entity.dto;
-
-public record UserInfoDto2(String userName, String univName, String regionName, String imgName ,String userTitle) {
-
-  public static UserInfoDto2 of (String userName, String univName, String regionName, String imgName ,String userTitle){
-    return new UserInfoDto2(userName, univName, regionName, imgName, userTitle);
-  }
-}
